Bug hot fix on retrieving ingredient from Firestore and editing quantity of the Ingredient in the dialog.

+++ app/src/main/java/com/example/nosh/fragments/ingredients/IngredientsFragment.java
@@ -118,7 +118,7 @@ public class IngredientsFragment extends Fragment implements Observer {
                     controller.update(
                             result.getString("hashcode"),
                             (Date) result.getSerializable("date"),
-                            result.getInt("qty"),
+                            result.getLong("qty"),
                             result.getDouble("unit"),
                             result.getString("name"),
                             result.getString("description"),
